More fixing up to use tensor types
